Use StrategicMergePatch to apply overrides
This commit removes the custom etcdaenixiov1alpha1.PodSpec type in
favor of corev1.PodSpec. As a result, StrategicMergePatch may be used
to correctly applythe overrides from EtcdCluster.Spec.PodTemplate.Spec
on top of the default PodSpec, generated by the operator. In turn, this
allows removal of dozens of lines of code responsible for custom merging
logic.